TibSTR-Former/requirements.txt

albumentations==1.4.19
docx>=0.2.4
lightning>=2.5.1.post0
line-profiler>=4.2.0
matplotlib>=3.10.1
numpy>=2.2.4
opencv-python>=4.11.0.86
pandas>=2.2.3
pillow>=11.1.0
tee>=0.0.3
torch>=2.7.0
torchvision>=0.22.0
tqdm>=4.67.1
